Add support for multiple virtualenvs
The original stacktest implementation required py27 to be universal
and ran everything out of it. That worked for nova at least for quite
a while, but now that there are things in the functional venv that
are not in py27, stacktest no longer works for functional tests.

This change does several things:

 - Adds stacktest-venv-pattern, which is a list of rules for determining what
   venv to use based on the filename of the buffer. The default rules retain
   the previous behavior of assuming py27 if nothing else matches. Any
   project with a nova-like tests/unit/ and tests/functional/ structure will
   automatically get py27 and functional venv usage respectively. Additional
   regexp rules can be added to make further tweaks.
 - Adds stacktest-venvify, which will determine the proper virtualenv using
   the pattern rules, and substitute $VENV in a command path with the proper
   value.
 - Changes the defaults for the subunit and testools commands to include a
   "$VENV" string instead of py27. Existing configurations that defined these
   as something specific will continue to work.

This works for me locally and via tramp in nova unit and functional tests.